Made GameCanvas abstract
